backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
psycopg2-binary>=2.9
pandas>=2.0
scikit-learn>=1.2
joblib>=1.2
numpy>=1.23
gunicorn>=20.1.0
python-dotenv>=1.0
